% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addEdge}
\alias{addEdge}
\title{addEdge}
\usage{
addEdge(tree, parent, child, type, rate)
}
\arguments{
\item{tree}{DiffTree to add transition to}

\item{parent}{parent population name ("from" population)}

\item{child}{child population name ("to" population)}

\item{type}{transition or event type ("alpha", "beta", "gamma1", "gamma2", "delta", "zeta", "mutation")}

\item{rate}{expected number of events per cell per time unit}
}
\description{
Used to designate transition between populations
}
\examples{
\dontrun{
addEdge(myTree, "Population A", "Population A", "alpha", 0.01)
addEdge(myTree, "Population A", "Population B", "gamma1", 0.05)
addEdge(myTree, "Population A", "Population A", "mutation", 1e-7)
}
}
