% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{allToCSV}
\alias{allToCSV}
\title{allToCSV}
\usage{
allToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of types file}

\item{prefixes}{a list of prefixes to convert}
}
\description{
convert .all file to CSV
}
