% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DiffTriangle}
\alias{DiffTriangle}
\title{DiffTriangle}
\usage{
DiffTriangle(tree, name, height, first_level = 1, label = 0)
}
\arguments{
\item{tree}{DiffTree object to add DiffTriangle to}

\item{name}{population name}

\item{height}{number of levels in structure, corresponding to number of levels until full maturation}

\item{first_level}{number of triangles to be stacked side-by-side or the number of cells at the first level of maturation}

\item{label}{initial probabilty that a cell receives a unique barcode}
}
\description{
Designates a DiffTriangle, a fixed-size structure used to model a population undergoing different levels of maturation
}
\details{
A DiffTriangle cell type is used to represent the downstream fully differentiated cells.  Cells are arranged in a triangle formation.  Cells enter the population on the highest level of the triangle, experiencing further differentiation and division to progress down the triangle.  When a new cell enters the DiffTriangle population, it causes an already existing cell on the highest level to divide and further differentiate to the next level of the triangle.  Those two cells each displace a pre-existing cell, causing them to divide and differentiate (thus generating four newly displaced cells), which in turn displace cells that further displace cells until reaching the lowest level of the triangle.  A displaced cell from the last row of the triangle can either be passed on to an offspring cell type (if there are further cell types in the hierarchy), or die out.  Importantly, DiffTriangle structures will not initiate any cellular events of their own, as differentiation waves throughout a triangle is only initiated when receiving a new cell from an upstream population.
The population size of a DiffTriangle is specified by two parameters; the first, z is the number of cell divisions until full maturation or the number of levels in the triangle.  The second is called the mfactor, which is the number of triangles to be stacked side-by-side.  If mfactor is greater than 1, then a cell entering the DiffTriangle population simply chooses at random which specific triangle to enter.
}
\examples{
\dontrun{
DiffTriangle(myTree, "Differentiated Population A", 12, 1000, 0.5)
}
}
