% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{eventsToCSV}
\alias{eventsToCSV}
\title{eventsToCSV}
\usage{
eventsToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of events file}

\item{prefixes}{a list of prefixes to convert}

\item{size}{initial population size}

\item{label}{initial probabilty that a cell receives a unique barcode}
}
\description{
convert .events file to CSV
}
