% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GrowingPop}
\alias{GrowingPop}
\title{GrowingPop}
\usage{
GrowingPop(name, size, label = 0)
}
\arguments{
\item{name}{population name}

\item{size}{initial population size}

\item{label}{initial probabilty that a cell receives a unique barcode}
}
\description{
Designates a GrowingPop, a structure used to simulate differentiation via a branching process
}
\examples{
\dontrun{
GrowingPop("Population A", 1000, 0.5)
}
}
