% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{labelToCSV}
\alias{labelToCSV}
\title{labelToCSV}
\usage{
labelToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of label file}

\item{prefixes}{a list of prefixes to convert}
}
\description{
convert .label file to CSV
}
