% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{mutToCSV}
\alias{mutToCSV}
\title{mutToCSV}
\usage{
mutToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of mut file}

\item{prefixes}{a list of prefixes to convert}
}
\description{
convert mut file to CSV
}
