% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{outputToCSV}
\alias{outputToCSV}
\title{outputToCSV}
\usage{
outputToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of output files}

\item{prefixes}{a list of prefixes to convert}
}
\description{
convert all output files to CSV
}
