% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{params2ToCSV}
\alias{params2ToCSV}
\title{params2ToCSV}
\usage{
params2ToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of params2 file}

\item{prefixes}{a list of prefixes to convert}
}
\description{
convert params2 file to CSV
}
