% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{paramsToCSV}
\alias{paramsToCSV}
\title{paramsToCSV}
\usage{
paramsToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of params file}

\item{prefixes}{a list of prefixes to convert}
}
\description{
convert params file to CSV
}
