% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{popToCSV}
\alias{popToCSV}
\title{popToCSV}
\usage{
popToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of pop file}

\item{prefixes}{a list of prefixes to convert}
}
\description{
convert .pop file to CSV
}
