% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readFileR}
\alias{readFileR}
\title{readFile}
\usage{
readFileR(indir, fname)
}
\arguments{
\item{indir}{input directory for tree files}

\item{filename}{output directory}
}
\value{
Nodelist nodelist represented by file
}
\description{
Runs simulation of a tree containing Fixed Population Size objects
[FixedPops & DiffTriangles].
}
