% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{sdiToCSV}
\alias{sdiToCSV}
\title{sdiToCSV}
\usage{
sdiToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of sdi file}

\item{prefixes}{a list of prefixes to convert}
}
\description{
convert sdi file to CSV
}
