% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulateTree}
\alias{simulateTree}
\title{simulateTree}
\usage{
simulateTree(tree, fixed = FALSE, time = 100, census = -1, indir = ".",
  outdir = ".", seed = NULL)
}
\arguments{
\item{tree}{DiffTree object to simulate}

\item{fixed}{boolean.  TRUE if simulating a fixed population hierarchy using FixedPops and DiffTriangles, FALSE if simulating a dynamic-sized population hierarchy using GrowingPops}

\item{time}{number of time units to simulate}

\item{census}{how often to output entire system population to filesystem}

\item{indir}{input directory for tree files}

\item{outdir}{output directory}

\item{seed}{numeric seed for random number generator}
}
\description{
Used to plot hierarchy structure using igraph
}
\examples{
\dontrun{
simulateTree(myTree, fixed = TRUE, time = 500, census = 1, indir = "input_dir", outdir = "output_dir")
}
}
