% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulateTreeCodeNew}
\alias{simulateTreeCodeNew}
\title{simulateTreeCodeNew}
\usage{
simulateTreeCodeNew(nObs = 10L, traverseFrequency = -1L, indir = "./",
  outdir = "./", seed = NULL)
}
\arguments{
\item{nObs}{number of observations for simulation time}

\item{traverseFrequency}{how often to output entire system population to filesystem}

\item{indir}{input directory for tree files}

\item{outdir}{output directory}

\item{seed}{numeric seed for random number generator}
}
\description{
Runs simulation of a tree structure using Gillespie algorithm
}
