% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransforms.R
\name{typesToCSV}
\alias{typesToCSV}
\title{typesToCSV}
\usage{
typesToCSV(directory = ".", prefixes = NULL)
}
\arguments{
\item{directory}{directory of types file}

\item{prefixes}{a list of prefixes to convert}
}
\description{
convert .types file to CSV
}
