% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DiffTriangle}
\alias{DiffTriangle}
\title{DiffTriangle}
\usage{
DiffTriangle(name, height, first_level = 1, label = 0)
}
\arguments{
\item{name}{population name}

\item{height}{number of levels in structure, corresponding to number of levels until full maturation}

\item{first_level}{number of triangles to be stacked side-by-side or the number of cells at the first level of maturation}

\item{label}{initial probabilty that a cell receives a unique barcode}
}
\description{
Designates a DiffTriangle, a fixed-size structure used to model a population undergoing different levels of maturation
}
\examples{
\dontrun{
DiffTriangle("Differentiated Population A", 12, 1000, 0.5)
}
}
