% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FixedPop}
\alias{FixedPop}
\title{FixedPop}
\usage{
FixedPop(tree, name, size, label = 0)
}
\arguments{
\item{tree}{DiffTree object to add FixedPop to}

\item{name}{population name}

\item{size}{initial population size}

\item{label}{initial probabilty that a cell receives a unique barcode}
}
\description{
Designates a FixedPop, a structure whose size remains constant throughout the course of a simulation
}
\examples{
\dontrun{
FixedPop(myTree, "Population A", 1000, 0.5)
}
}
