% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotTree}
\alias{plotTree}
\title{plotTree}
\usage{
plotTree(allEdge = F)
}
\arguments{
\item{allEdge}{boolean variable set to TRUE to plot all transitions, set to FALSE to only show gamma1, gamma2, and beta edges}
}
\description{
Used to plot hierarchy structure using igraph
}
\examples{
\dontrun{
plotTree(allEdge = TRUE)
}
}
