% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setRoot}
\alias{setRoot}
\title{setRoot}
\usage{
setRoot(popName)
}
\arguments{
\item{popName}{name of furthest upstream population}
}
\description{
Used to designate furthest upstream population in the hierarchy
}
