% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writeTree}
\alias{writeTree}
\title{writeTree}
\usage{
writeTree(outdir = "/graph_files/")
}
\arguments{
\item{outdir}{output directory for files}
}
\description{
Used to create tree files needed by C++ simulation
}
\examples{
\dontrun{
writeTree(outdir = "C:/diffpop/example1/")
}
}
